+++ doc/changes.rst
@@ -6,7 +6,11 @@ Changes
 New in 0.34
 -----------
 
-* Added: :func:`normalize_in_cells`.
+* Added:
+
+ - :func:`normalize_in_cells`.
+ - :func:`boosting`: Option to store TRFs for the different test partitions
+   (``partition_results``).
 
 
 New in 0.33